+++ megaradrp/core/recipe.py
@@ -23,7 +23,7 @@ from astropy.io import fits
 
 import numina.array.combine as combine
 from numina.flow import SerialFlow
-from numina.flow.processing import BiasCorrector, BadPixelCorrector
+from numina.flow.processing import BiasCorrector
 from numina.core import BaseRecipe
 from numina.core.dataholders import Product
 from numina.core.products import QualityControlProduct
@@ -38,28 +38,24 @@ class MegaraBaseRecipe(BaseRecipe):
     qc = Product(QualityControlProduct, dest='qc')
 
     def __init__(self, version):
-        self.__flow = {'ArcCalibrationRecipe':[OverscanCorrector, TrimImage, BiasCorrector, BadPixelCorrector],
-                       'BadPixelsMaskRecipe':[OverscanCorrector, TrimImage, BiasCorrector],
-                       'BiasRecipe':[OverscanCorrector, TrimImage, BadPixelCorrector],
-                       'FiberFlatRecipe':[OverscanCorrector, TrimImage, BiasCorrector, BadPixelCorrector],
+        self.__flow = {'BiasRecipe':[OverscanCorrector, TrimImage],
+                       'ArcCalibrationRecipe':[OverscanCorrector, TrimImage, BiasCorrector],
+                       # 'ArcCalibrationRecipe':[BadPixelsMaskCorrector,OverscanCorrector, TrimImage, BiasCorrector],
+                       'FiberFlatRecipe':[OverscanCorrector, TrimImage, BiasCorrector],
                        'TraceMapRecipe':[OverscanCorrector, TrimImage, BiasCorrector],
-
+                       'BadPixelsMaskRecipe':[OverscanCorrector, TrimImage, BiasCorrector],
                        }
         super(MegaraBaseRecipe, self).__init__(version=version)
 
     def __generate_flow(self, params):
         import copy
-        ff = self.__flow[self.__class__.__name__]
+        ff  = self.__flow[self.__class__.__name__]
         flow = copy.deepcopy(ff)
+        # flow = self.__flow[self.__class__.__name__]
         try:
             for cont in range(len(flow)):
                 if issubclass(BiasCorrector, flow[cont]):
                     flow[cont] = (flow[cont](params['biasmap']))
-                elif issubclass(BadPixelCorrector,flow[cont]):
-                    if 'bpm' in params.keys():
-                        flow[cont] = (flow[cont](params['bpm']))
-                    else:
-                        del (flow[cont])
                 elif issubclass(TrimImage, flow[cont]) or issubclass(OverscanCorrector, flow[cont]):
                     flow[cont] = (flow[cont]())
             basicflow = SerialFlow(flow)
@@ -70,14 +66,14 @@ class MegaraBaseRecipe(BaseRecipe):
         del flow
         return basicflow
 
-    def bias_process_common(self, obresult, img):
+    def bias_process_common(self, obresult, master_bias):
 
-        # with master_bias.open() as hdul:
-        #     mbias = hdul[0].data.copy()
+        with master_bias.open() as hdul:
+            mbias = hdul[0].data.copy()
 
-        hdu, data = self.hdu_creation(obresult, img)
+        hdu, data = self.hdu_creation(obresult, {'biasmap':mbias})
 
-        hdr = hdu[0].header
+        hdr = hdu.header
         # FIXME: this is incorrect in general
         hdr['IMGTYP'] = ('FIBER_FLAT', 'Image type')
         hdr['NUMTYP'] = ('MASTER_FIBER_FLAT', 'Data product type')
@@ -86,13 +82,13 @@ class MegaraBaseRecipe(BaseRecipe):
 
         varhdu = fits.ImageHDU(data[1], name='VARIANCE')
         num = fits.ImageHDU(data[2], name='MAP')
-        reduced = fits.HDUList(hdu + [varhdu, num])
+        reduced = fits.HDUList([hdu, varhdu, num])
         return reduced
 
-    def hdu_creation(self, obresult, params={}):
+    def hdu_creation(self, obresult, params=None):
 
         basicflow = self.__generate_flow(params)
-        lista = []
+
         cdata = []
         try:
             for frame in obresult.images:
@@ -105,12 +101,8 @@ class MegaraBaseRecipe(BaseRecipe):
             data = combine.median([d[0].data for d in cdata], dtype='float32')
             template_header = cdata[0][0].header
             hdu = fits.PrimaryHDU(data[0], header=template_header)
-            lista.append(hdu)
-            # if has_mask:
-            #     hdumask = fits.ImageHDU(themask)
-            #     lista.append(hdumask)
         finally:
             for hdulist in cdata:
                 hdulist.close()
 
-        return fits.HDUList(lista), data
+        return hdu, data
