+++ megaradrp/products.py
@@ -59,8 +59,6 @@ class MasterDark(MEGARAProductFrame):
 class MasterFiberFlat(MEGARAProductFrame):
     pass
 
-class MasterBPM(MEGARAProductFrame):
-    pass
 
 class MasterSensitivity(MEGARAProductFrame):
     pass
