+++ megaradrp/recipes/calibration/arc.py
@@ -42,7 +42,7 @@ from numina.array.peaks.findpeaks1D import refinePeaks_spectrum
 
 from megaradrp.core.recipe import MegaraBaseRecipe
 from megaradrp.products import TraceMap
-from megaradrp.requirements import MasterBiasRequirement, MasterBPMRequirement
+from megaradrp.requirements import MasterBiasRequirement
 from megaradrp.core.processing import apextract_tracemap
 
 _logger = logging.getLogger('numina.recipes.megara')
@@ -54,7 +54,6 @@ class ArcCalibrationRecipe(MegaraBaseRecipe):
     # Requirements
     obresult = ObservationResultRequirement()
     master_bias = MasterBiasRequirement()
-    master_bpm = MasterBPMRequirement()
     tracemap = Requirement(TraceMap, 'Trace information of the Apertures')
     lines_catalog = Requirement(LinesCatalog, 'Catalog of lines')
     polynomial_degree = Parameter(2, 'Polynomial degree of arc calibration')
@@ -68,11 +67,7 @@ class ArcCalibrationRecipe(MegaraBaseRecipe):
 
     def run(self, rinput):
         # Basic processing
-        with rinput.master_bias.open() as hdul:
-            mbias = hdul[0].data.copy()
-
-        reduced = self.bias_process_common(rinput.obresult, {'biasmap':mbias,
-                                                             'bpm':None})
+        reduced = self.bias_process_common(rinput.obresult, rinput.master_bias)
 
         _logger.info('extract fibers')
         _logger.info('extract fibers, %i', len(rinput.tracemap))
