+++ megaradrp/recipes/calibration/bias.py
@@ -26,15 +26,14 @@ from numina.core.requirements import ObservationResultRequirement
 
 from megaradrp.core.recipe import MegaraBaseRecipe
 from megaradrp.products import MasterBias
-from megaradrp.requirements import MasterBPMRequirement
 
 _logger = logging.getLogger('numina.recipes.megara')
 
 
 class BiasRecipe(MegaraBaseRecipe):
     """Process BIAS images and create MASTER_BIAS."""
-    obresult = ObservationResultRequirement()
 
+    obresult = ObservationResultRequirement()
     biasframe = Product(MasterBias)
 
     def __init__(self):
@@ -49,7 +48,7 @@ class BiasRecipe(MegaraBaseRecipe):
 
         hdu, data = self.hdu_creation(rinput.obresult)
 
-        hdr = hdu[0].header
+        hdr = hdu.header
         hdr = self.set_base_headers(hdr)
         hdr['IMGTYP'] = ('BIAS', 'Image type')
         hdr['NUMTYP'] = ('MASTER_BIAS', 'Data product type')
@@ -57,7 +56,7 @@ class BiasRecipe(MegaraBaseRecipe):
 
         varhdu = fits.ImageHDU(data[1], name='VARIANCE')
         num = fits.ImageHDU(data[2], name='MAP')
-        hdulist = fits.HDUList(hdu + [varhdu, num])
+        hdulist = fits.HDUList([hdu, varhdu, num])
         _logger.info('bias reduction ended')
 
         result = self.create_result(biasframe=hdulist)
