+++ megaradrp/recipes/calibration/trace.py
@@ -30,9 +30,10 @@ from numina.array.trace.traces import trace
 from numina.core import Product
 from numina.core.requirements import ObservationResultRequirement
 
+from megaradrp.core.processing import apextract_tracemap
 from megaradrp.products import MasterFiberFlat, TraceMap
 from megaradrp.core.recipe import MegaraBaseRecipe
-from megaradrp.requirements import MasterBiasRequirement, MasterBPMRequirement
+from megaradrp.requirements import MasterBiasRequirement
 from megaradrp.trace.traces import init_traces
 
 _logger = logging.getLogger('numina.recipes.megara')
@@ -41,7 +42,6 @@ _logger = logging.getLogger('numina.recipes.megara')
 class TraceMapRecipe(MegaraBaseRecipe):
 
     obresult = ObservationResultRequirement()
-    master_bpm = MasterBPMRequirement()
     master_bias = MasterBiasRequirement()
     fiberflat_frame = Product(MasterFiberFlat)
     traces = Product(TraceMap)
@@ -54,11 +54,7 @@ class TraceMapRecipe(MegaraBaseRecipe):
     def run(self, rinput):
 
         # Basic processing
-        with rinput.master_bias.open() as hdul:
-            mbias = hdul[0].data.copy()
-
-        reduced = self.bias_process_common(rinput.obresult, {'biasmap':mbias})
-
+        reduced = self.bias_process_common(rinput.obresult, rinput.master_bias)
         data = reduced[0].data
         cstart = 2000
         hs = 3
